+++ covasim/interventions.py
@@ -668,8 +668,7 @@ class clip_edges(Intervention):
     def finalize(self, sim):
         ''' Ensure the edges get deleted at the end '''
         super().finalize()
-        if sim.t == sim.tvec[-1]:
-            self.contacts = None # Reset to save memory
+        self.contacts = None # Reset to save memory
         return
 
 
@@ -774,6 +773,13 @@ class test_num(Intervention):
         self.daily_tests = process_daily_data(self.daily_tests, sim, self.start_day)
         self.ili_prev    = process_daily_data(self.ili_prev,    sim, self.start_day)
 
+        return
+
+
+    def finalize(self, sim):
+        ''' Ensure variables with large memory footprints get erased '''
+        super().finalize()
+        self.subtarget = None # Reset to save memory
         return
 
 
@@ -900,6 +906,13 @@ class test_prob(Intervention):
         self.end_day   = preprocess_day(self.end_day,   sim)
         self.days      = [self.start_day, self.end_day]
         self.ili_prev  = process_daily_data(self.ili_prev, sim, self.start_day)
+        return
+
+
+    def finalize(self, sim):
+        ''' Ensure variables with large memory footprints get erased '''
+        super().finalize()
+        self.subtarget = None # Reset to save memory
         return
 
 
@@ -1142,7 +1155,7 @@ class simple_vaccine(Intervention):
     relative susceptibility and the probability of developing symptoms if still
     infected, this intervention stores several types of data:
 
-        - ``vaccinations``:      the number of vaccine doses per person
+        - ``doses``:      the number of vaccine doses per person
         - ``vaccination_dates``: list of dates per person
         - ``orig_rel_sus``:      relative susceptibility per person at the beginning of the simulation
         - ``orig_symp_prob``:    probability of developing symptoms per person at the beginning of the simulation
@@ -1182,10 +1195,10 @@ class simple_vaccine(Intervention):
 
 
     def initialize(self, sim):
-        ''' Fix the dates and store the vaccinations '''
+        ''' Fix the dates and store the doses '''
         super().initialize()
         self.days = process_days(sim, self.days)
-        self.vaccinations      = np.zeros(sim.n, dtype=cvd.default_int) # Number of doses given per person
+        self.doses             = np.zeros(sim.n, dtype=cvd.default_int) # Number of doses given per person of this vaccine
         self.vaccination_dates = [[]] * sim.n # Store the dates when people are vaccinated
         self.orig_rel_sus      = sc.dcp(sim.people.rel_sus) # Keep a copy of pre-vaccination susceptibility
         self.orig_symp_prob    = sc.dcp(sim.people.symp_prob) # ...and symptom probability
@@ -1209,7 +1222,7 @@ class simple_vaccine(Intervention):
             vacc_inds = cvu.true(cvu.binomial_arr(vacc_probs)) # Calculate who actually gets vaccinated
 
             # Calculate the effect per person
-            vacc_doses = self.vaccinations[vacc_inds] # Calculate current doses
+            vacc_doses = self.doses[vacc_inds] # Calculate current doses
             eff_doses = np.minimum(vacc_doses, len(self.cumulative)-1) # Convert to a valid index
             vacc_eff = self.cumulative[eff_doses] # Pull out corresponding effect sizes
             rel_sus_eff  = (1.0 - vacc_eff) + vacc_eff*self.rel_sus
@@ -1222,13 +1235,13 @@ class simple_vaccine(Intervention):
             # Update counters
             self.mod_rel_sus[vacc_inds]   *= rel_sus_eff
             self.mod_symp_prob[vacc_inds] *= rel_symp_eff
-            self.vaccinations[vacc_inds] += 1
+            self.doses[vacc_inds] += 1
             for v_ind in vacc_inds:
                 self.vaccination_dates[v_ind].append(sim.t)
 
             # Update vaccine attributes in sim
             sim.people.vaccinated[vacc_inds] = True
-            sim.people.vaccinations[vacc_inds] += 1
+            sim.people.doses[vacc_inds] += 1
 
         return
 
@@ -1252,12 +1265,13 @@ class BaseVaccination(Intervention):
     These are:
 
         - ``vaccinated``:        whether or not a person is vaccinated
-        - ``vaccinations``:      the number of vaccine doses per person
-        - ``vaccination_dates``: list of vaccination dates per person
+        - ``doses``:             the number of vaccine doses per person
+        - ``vaccination_dates``: integer; date of last dose
 
     Args:
         vaccine (dict/str): which vaccine to use; see below for dict parameters
         label        (str): if vaccine is supplied as a dict, the name of the vaccine
+        booster      (boolean): whether the vaccine is a booster, i.e. whether vaccinated people are eligible
         kwargs     (dict): passed to Intervention()
 
     If ``vaccine`` is supplied as a dictionary, it must have the following parameters:
@@ -1269,8 +1283,8 @@ class BaseVaccination(Intervention):
         - ``target_eff``: the target efficacy from which to calculate initial antibody and boosting.
         must be supplied as a list, where length of list is equal to number of doses
         - ``nab_eff``:   the waning efficacy of neutralizing antibodies at preventing infection
-        - ``doses``:     the number of doses required to be fully vaccinated
-        - ``interval``:  the interval between doses
+        - ``doses``:     the number of doses required to be fully vaccinated with this vaccine
+        - ``interval``:  the interval between doses (integer)
         - entries for efficacy against each of the variants (e.g. ``b117``)
 
     See ``parameters.py`` for additional examples of these parameters.
@@ -1282,7 +1296,7 @@ class BaseVaccination(Intervention):
         self.index     = None # Index of the vaccine in the sim; set later
         self.label     = label # Vaccine label (used as a dict key)
         self.p         = None # Vaccine parameters
-        self.vaccinations = None # Record the number of doses given per person *by this intervention*
+        self.doses = None # Record the number of doses given per person *by this intervention*
         self.vaccination_dates = None # Store the dates that each person was last vaccinated *by this intervention*
 
         self._parse_vaccine_pars(vaccine=vaccine) # Populate
@@ -1377,10 +1391,10 @@ class BaseVaccination(Intervention):
                     boost = (2**boosted_nab)/(2**peak_nab)
                     self.p['nab_boost'] = boost
             else:
-                errormsg = f'Provided mismatching efficacies and doses.'
+                errormsg = 'Provided mismatching efficacies and doses.'
                 raise ValueError(errormsg)
 
-        self.vaccinations         = np.zeros(sim['pop_size'], dtype=cvd.default_int) # Number of doses given per person
+        self.doses         = np.zeros(sim['pop_size'], dtype=cvd.default_int) # Number of doses given per person
         self.vaccination_dates    = [[]] * sim.n # Store the dates when people are vaccinated
 
         sim['vaccine_pars'][self.label] = self.p # Store the parameters
@@ -1389,17 +1403,21 @@ class BaseVaccination(Intervention):
 
         return
 
+
+    def finalize(self, sim):
+        ''' Ensure variables with large memory footprints get erased '''
+        super().finalize()
+        self.subtarget = None # Reset to save memory
+        return
+
+
     def select_people(self, sim):
         """
         Return an array of indices of people to vaccinate
-
         Derived classes must implement this function to determine who to vaccinate at each timestep
-
         Args:
             sim: A cv.Sim instance
-
         Returns: Array of person indices
-
         """
         raise NotImplementedError
 
@@ -1409,8 +1427,8 @@ class BaseVaccination(Intervention):
         Vaccinate people
 
         This method applies the vaccine to the requested people indices. The indices of people vaccinated
-        is returned. These may be different to the requested indices, because anyone that is dead or who is
-        already fully vaccinated (according to the dose schedule for this vaccine) will be skipped. This could
+        is returned. These may be different to the requested indices, because anyone that is dead will be
+        skipped, as well as anyone already fully vaccinated (if booster=False). This could
         occur if a derived class does not filter out such people in its `select_people` method.
 
         Args:
@@ -1426,26 +1444,33 @@ class BaseVaccination(Intervention):
         else:
             assert t <= sim.t, 'Overriding the vaccination day should only be used for historical vaccination' # High potential for errors to creep in if future vaccines could be scheduled here
 
+        # Perform checks
         vacc_inds = vacc_inds[~sim.people.dead[vacc_inds]] # Skip anyone that is dead
-        vacc_inds = vacc_inds[self.vaccinations[vacc_inds] < self.p['doses']] # Skip anyone that is already fully vaccinated. Otherwise, they will receive the 2nd dose boost cumulatively for every subsequent dose
+        # Skip anyone that has already had all the doses of *this* vaccine (not counting boosters).
+        # Otherwise, they will receive the 2nd dose boost cumulatively for every subsequent dose.
+        # Note, this does not preclude someone from getting additional doses of another vaccine (e.g. a booster)
+        vacc_inds = vacc_inds[self.doses[vacc_inds] < self.p['doses']]
+
+        # Extract indices of already-vaccinated people and get indices of newly-vaccinated
+        prior_vacc = cvu.true(sim.people.vaccinated)
+        new_vacc   = np.setdiff1d(vacc_inds, prior_vacc)
 
         if len(vacc_inds):
-            self.vaccinations[vacc_inds] += 1
+            self.doses[vacc_inds] += 1
             for v_ind in vacc_inds:
                 self.vaccination_dates[v_ind].append(sim.t)
 
             sim.people.vaccinated[vacc_inds] = True
             sim.people.vaccine_source[vacc_inds] = self.index
-            sim.people.vaccinations[vacc_inds] += 1
+            sim.people.doses[vacc_inds] += 1
             sim.people.date_vaccinated[vacc_inds] = t
 
             cvi.update_peak_nab(sim.people, vacc_inds, nab_pars=self.p, natural=False)
 
             if t >= 0:
                 # Only record these quantities by default if it's not a historical dose
-                sim.people.flows['new_vaccinations'] += len(vacc_inds) # Count number of doses given
-                sim.people.flows['new_vaccinated']   += np.count_nonzero(sim.people.vaccinations[vacc_inds] == 1) # Count number of people not already vaccinated given doses
-
+                sim.people.flows['new_doses'] += len(vacc_inds) # Count number of doses given
+                sim.people.flows['new_vaccinated']   += len(new_vacc) # Count number of people not already vaccinated given doses
 
         return vacc_inds
 
@@ -1463,7 +1488,7 @@ class BaseVaccination(Intervention):
         ''' Shrink vaccination intervention '''
         obj = super().shrink(in_place=in_place)
         obj.vaccinated = None
-        obj.vaccinations = None
+        obj.doses = None
         obj.vaccination_dates = None
         if hasattr(obj, 'second_dose_days'):
             obj.second_dose_days = None
@@ -1486,6 +1511,27 @@ def vaccinate(*args, **kwargs):
         return vaccinate_prob(*args, **kwargs)
 
 
+def check_doses(doses, interval):
+    '''Check doses and intervals'''
+
+    # First check that they're both numbers
+    if not sc.checktype(doses, int):
+        raise ValueError(f'Doses must be an integer, not {doses}.')
+    if interval is not None and not sc.isnumber(interval):
+        errormsg = f"Can't understand the dosing interval given by '{interval}'. Dosing interval should be a number."
+        raise ValueError(errormsg)
+
+    # Now check that they're compatible
+    if doses == 1 and interval is not None:
+        raise ValueError("Can't use dosing intervals for vaccines with only one dose.")
+    elif doses == 2 and interval is None:
+        raise ValueError('Must specify a dosing interval if using a vaccine with more than one dose.')
+    elif doses > 2:
+        raise NotImplementedError('Scheduling three or more doses not yet supported')
+
+    return
+
+
 class vaccinate_prob(BaseVaccination):
     '''
     Probability-based vaccination
@@ -1498,6 +1544,7 @@ class vaccinate_prob(BaseVaccination):
         label        (str): if vaccine is supplied as a dict, the name of the vaccine
         days     (int/arr): the day or array of days to apply the interventions
         prob       (float): probability of being vaccinated (i.e., fraction of the population)
+        booster    (bool): whether it's a booster (i.e. targeted to vaccinated people) or not
         subtarget  (dict): subtarget intervention to people with particular indices (see test_num() for details)
         kwargs     (dict): passed to Intervention()
 
@@ -1507,7 +1554,7 @@ class vaccinate_prob(BaseVaccination):
         - ``nab_init``:  the initial antibody level (higher = more protection)
         - ``nab_boost``: how much of a boost being vaccinated on top of a previous dose or natural infection provides
         - ``doses``:     the number of doses required to be fully vaccinated
-        - ``interval``:  the interval between doses
+        - ``interval``:  the interval between doses (integer)
         - entries for efficacy against each of the strains (e.g. ``b117``)
 
     See ``parameters.py`` for additional examples of these parameters.
@@ -1517,11 +1564,13 @@ class vaccinate_prob(BaseVaccination):
         pfizer = cv.vaccinate_prob(vaccine='pfizer', days=30, prob=0.7)
         cv.Sim(interventions=pfizer, use_waning=True).run().plot()
     '''
-    def __init__(self, vaccine, days, label=None, prob=1.0, subtarget=None, **kwargs):
+    def __init__(self, vaccine, days, label=None, prob=1.0, subtarget=None, booster=False, **kwargs):
         super().__init__(vaccine,label=label,**kwargs) # Initialize the Intervention object
         self.days      = sc.dcp(days)
         self.prob      = prob
+        self.booster   = booster
         self.subtarget = subtarget
+        self.booster   = booster
         self.second_dose_days = None  # Track scheduled second doses
         return
 
@@ -1529,6 +1578,7 @@ class vaccinate_prob(BaseVaccination):
         super().initialize(sim)
         self.days = process_days(sim, self.days) # days that group becomes eligible
         self.second_dose_days     = [None]*sim.npts # People who get second dose (if relevant)
+        check_doses(self.p['doses'], self.p['interval'])
         return
 
     def select_people(self, sim):
@@ -1540,20 +1590,28 @@ class vaccinate_prob(BaseVaccination):
             # Vaccinate people with their first dose
             for _ in find_day(self.days, sim.t, interv=self, sim=sim):
                 vacc_probs = np.zeros(sim['pop_size'])
-                unvacc_inds = sc.findinds(~sim.people.vaccinated)
+
+                # Find eligible people
+                vacc_probs[cvu.true(sim.people.dead)] *= 0.0  # Do not vaccinate dead people
+                # Eligibility depends on whether it's a booster or not
+                # If this is a booster, exclude unvaccinated people; otherwise, exclude vaccinated people
+                if self.booster:    eligible_inds = sc.findinds(sim.people.vaccinated)
+                else:               eligible_inds = sc.findinds(~sim.people.vaccinated)
+                vacc_probs[eligible_inds] = self.prob  # Assign equal vaccination probability to everyone
+
+                # Apply any subtargeting
                 if self.subtarget is not None:
                     subtarget_inds, subtarget_vals = get_subtargets(self.subtarget, sim)
                     vacc_probs[subtarget_inds] = subtarget_vals  # People being explicitly subtargeted
-                else:
-                    vacc_probs[unvacc_inds] = self.prob  # Assign equal vaccination probability to everyone
-                vacc_probs[cvu.true(sim.people.dead)] *= 0.0  # Do not vaccinate dead people
+
                 vacc_inds = cvu.true(cvu.binomial_arr(vacc_probs))  # Calculate who actually gets vaccinated
 
                 if len(vacc_inds):
                     if self.p.interval is not None:
-                        next_dose_day = sim.t + self.p.interval
-                        if next_dose_day < sim['n_days']:
-                            self.second_dose_days[next_dose_day] = vacc_inds
+                        # Schedule the doses
+                        next_dose_days = sim.t + self.p.interval
+                        if next_dose_days < sim['n_days']:
+                            self.second_dose_days[next_dose_days] = vacc_inds
 
             # Also, if appropriate, vaccinate people with their second dose
             vacc_inds_dose2 = self.second_dose_days[sim.t]
@@ -1565,7 +1623,6 @@ class vaccinate_prob(BaseVaccination):
 
 def process_sequence(sequence, sim):
     ''' Handle different types of prioritization sequence for vaccination '''
-
     if callable(sequence):
         sequence = sequence(sim.people)
     elif sequence == 'age':
@@ -1577,14 +1634,24 @@ def process_sequence(sequence, sim):
     else:
         errormsg = f'Unable to interpret sequence {type(sequence)}: must be None, "age", callable, or an array'
         raise TypeError(errormsg)
-
     return sequence
+
+
+def process_doses(num_doses, sim):
+    ''' Handle different types of dose data'''
+    if sc.isnumber(num_doses):
+        num_people = num_doses
+    elif callable(num_doses):
+        num_people = num_doses(sim)
+    elif sim.t in num_doses:
+        num_people = num_doses[sim.t]
+    else:
+        num_people = 0
+    return num_people
 
 
 class vaccinate_num(BaseVaccination):
     '''
-    Sequence-based vaccination
-
     This vaccine intervention allocates vaccines in a pre-computed order of
     distribution, at a specified rate of doses per day. Second doses are prioritized
     each day.
@@ -1592,6 +1659,8 @@ class vaccinate_num(BaseVaccination):
     Args:
         vaccine (dict/str): which vaccine to use; see below for dict parameters
         label        (str): if vaccine is supplied as a dict, the name of the vaccine
+        booster    (bool): whether it's a booster (i.e. targeted to vaccinated people) or not
+        subtarget  (dict): subtarget intervention to people with particular indices (see test_num() for details)
         sequence: Specify the order in which people should get vaccinated. This can be
 
             - An array of person indices in order of vaccination priority
@@ -1611,31 +1680,29 @@ class vaccinate_num(BaseVaccination):
         **kwargs: Additional arguments passed to ``cv.BaseVaccination``
 
     **Example**::
-
-        def age_sequence(people): return np.argsort(-people.age)
-        pfizer = cv.vaccinate_num(vaccine='pfizer', sequence=age_sequence, num_doses=100)
+        pfizer = cv.vaccinate_num(vaccine='pfizer', sequence='age', num_doses=100)
         cv.Sim(interventions=pfizer, use_waning=True).run().plot()
     '''
-    def __init__(self, vaccine, num_doses, sequence=None, **kwargs):
+
+    def __init__(self, vaccine, num_doses, booster=False, subtarget=None, sequence=None, **kwargs):
         super().__init__(vaccine,**kwargs) # Initialize the Intervention object
-        self.sequence = sequence
-        self.num_doses = num_doses
-        self._scheduled_doses = defaultdict(set)  # Track scheduled second doses
+        self.sequence   = sequence
+        self.num_doses  = num_doses
+        self.booster    = booster
+        self.subtarget  = subtarget
+        self._scheduled_doses = defaultdict(set)  # Track scheduled second doses, where applicable
         return
 
 
     def initialize(self, sim):
+
         super().initialize(sim)
 
-        # Convert any dates to simulation days
-        if isinstance(self.num_doses, dict):
+        # Perform checks and process inputs
+        if isinstance(self.num_doses, dict): # Convert any dates to simulation days
             self.num_doses = {sim.day(k):v for k, v in self.num_doses.items()}
-
         self.sequence = process_sequence(self.sequence, sim)
-
-
-        if self.p['doses'] > 2:
-            raise NotImplementedError('Scheduling three or more doses not yet supported')
+        check_doses(self.p['doses'], self.p['interval'])
 
         return
 
@@ -1643,25 +1710,17 @@ class vaccinate_num(BaseVaccination):
     def select_people(self, sim):
 
         # Work out how many people to vaccinate today
-        if sc.isnumber(self.num_doses):
-            num_people = self.num_doses
-        elif callable(self.num_doses):
-            num_people = self.num_doses(sim)
-        elif sim.t in self.num_doses:
-            num_people = self.num_doses[sim.t]
-        else:
-            # If nobody gets vaccinated today, just return an empty list
-            return np.array([])
-
+        num_people = process_doses(self.num_doses, sim)
+        if num_people == 0: return np.array([])
         num_agents = int(np.round(num_people / sim["pop_scale"]))
 
-        # First, see how many scheduled doses we are going to deliver
+        # First, see how many scheduled second doses we are going to deliver
         if self._scheduled_doses[sim.t]:
             scheduled = np.fromiter(self._scheduled_doses[sim.t], dtype=cvd.default_int) # Everyone scheduled today
-            scheduled = scheduled[(sim.people.vaccinations[scheduled]<self.p['doses']) & ~sim.people.dead[scheduled]] # Remove fully vaccinated or dead
+            scheduled = scheduled[(self.doses[scheduled] < self.p['doses']) & ~sim.people.dead[scheduled]] # Remove anyone who's already had all doses of this vaccine, also dead people
 
             # If there are more people due for a second dose than there are doses, vaccinate as many second doses
-            # as possible, and add the remainder to tomorrow's vaccinations. At the moment, they don't get priority
+            # as possible, and add the remainder to tomorrow's doses. At the moment, they don't get priority
             # because the order of the scheduling doesn't matter (so there is a chance someone could go for several days
             # before being allocated their second dose) but then there is some flexibility in the dosing schedules anyway
             # e.g. Pfizer being 3-6 weeks in some jurisdictions
@@ -1672,13 +1731,26 @@ class vaccinate_num(BaseVaccination):
         else:
             scheduled = np.array([], dtype=cvd.default_int)
 
-        # Next, work out who is eligible for a first dose
-        # Anyone who has received at least one dose of a vaccine would have had subsequent doses scheduled
-        # and therefore should not be selected here
-        first_dose_eligible = self.sequence[~sim.people.vaccinated[self.sequence] & ~sim.people.dead[self.sequence]]
+        # Next, work out who is eligible for their first dose
+        vacc_probs = np.ones(sim.n)  # Begin by assigning equal weight (converted to a probability) to everyone
+        vacc_probs[cvu.true(sim.people.dead)] = 0.0  # Dead people are not eligible
+
+        # Apply any subtargeting for this vaccination
+        if self.subtarget is not None:
+            subtarget_inds, subtarget_vals = get_subtargets(self.subtarget, sim)
+            vacc_probs[subtarget_inds] = vacc_probs[subtarget_inds]*subtarget_vals
+
+        # If this is a booster, exclude unvaccinated people; otherwise, exclude vaccinated people
+        if self.booster:    vacc_probs[cvu.false(sim.people.vaccinated)] = 0.0
+        else:               vacc_probs[cvu.true(sim.people.vaccinated)]  = 0.0 # Anyone who's received at least one dose is counted as vaccinated
+
+        # All remaining people can be vaccinated, although anyone who has received half of a multi-dose
+        # vaccine would have had subsequent doses scheduled and therefore should not be selected here
+        first_dose_eligible = self.sequence[cvu.binomial_arr(vacc_probs[self.sequence])]
 
         if len(first_dose_eligible) == 0:
             return scheduled  # Just return anyone that is scheduled
+
         elif len(first_dose_eligible) > num_agents:
             # Truncate it to the number of agents for performance when checking whether anyone scheduled overlaps with first doses to allocate
             first_dose_eligible = first_dose_eligible[:num_agents] # This is the maximum number of people we could vaccinate this timestep, if there are no second doses allocated
@@ -1696,9 +1768,11 @@ class vaccinate_num(BaseVaccination):
         # Schedule subsequent doses
         # For vaccines with >2 doses, scheduled doses will also need to be checked
         if self.p['doses'] > 1:
-            self._scheduled_doses[sim.t+self.p['interval']].update(first_dose_inds)
+            self._scheduled_doses[sim.t+self.p.interval].update(first_dose_inds)
 
-        return np.concatenate([scheduled, first_dose_inds])
+        vacc_inds = np.concatenate([scheduled, first_dose_inds])
+
+        return vacc_inds
 
 
 
@@ -1807,13 +1881,13 @@ class historical_vaccinate_prob(BaseVaccination):
             inds = self.select_people(sim, t)
             if len(inds):
                 inds = self.vaccinate(sim, inds, t=t)
-                sim.results['new_vaccinations'][0] += len(inds)
-                sim.results['new_vaccinated'][0] += np.count_nonzero(sim.people.vaccinations[inds] == 1)
+                sim.results['new_doses'][0] += len(inds)
+                sim.results['new_vaccinated'][0] += np.count_nonzero(sim.people.doses[inds] == 1)
 
             # we need to update the NAbs as it is a cumulative effect
             # this will mess up those who are the seed infections if not reset to naive (see above)
             sim.people.t = t
-            to_update = cvu.true(self.vaccinations > 0)  # Update nabs for anyone vaccinated using this intervention
+            to_update = cvu.true(self.doses > 0)  # Update nabs for anyone vaccinated using this intervention
             if len(to_update):
                 cvi.update_nab(sim.people, inds=to_update)